refactor: split / remove unused code / DRY / pm2 enterprise|plus - breaking: .interact dropped
