+++ lib/API.js
@@ -390,7 +390,7 @@ class API {
     that.getVersion(function(err, new_version) {
       // If not linked to PM2 plus, and update PM2 to latest, display motd.update
       if (!that.gl_is_km_linked && !err && (pkg.version != new_version)) {
-        var dt = fs.readFileSync(path.join(__dirname, that._conf.KEYMETRICS_UPDATE));
+        var dt = fs.readFileSync(path.join(__dirname, that._conf.PM2_UPDATE));
         console.log(dt.toString());
       }
 
@@ -1671,10 +1671,14 @@ class API {
 //////////////////////////
 
 require('./API/Extra.js')(API);
-require('./API/Interaction.js')(API);
 require('./API/Deploy.js')(API);
 require('./API/Modules/Modules.js')(API);
-require('./API/PM2/PM2IO.js')(API);
+
+require('./API/pm2-plus/link.js')(API);
+require('./API/pm2-plus/process-selector.js')(API);
+require('./API/pm2-plus/helpers.js')(API);
+require('./API/pm2-plus/PM2IO.js')(API);
+
 require('./API/Configuration.js')(API);
 require('./API/Version.js')(API);
 require('./API/Startup.js')(API);
