+++ lib/API/pm2-plus/auth-strategies/CliAuth.js
@@ -13,7 +13,7 @@ const path = require('path')
 const os = require('os')
 const needle = require('needle');
 const chalk = require('chalk')
-const cst = require('../../../constants.js');
+const cst = require('../../../../constants.js');
 
 module.exports = class CustomStrategy extends AuthStrategy {
   // the client will try to call this but we handle this part ourselves
