+++ lib/API/pm2-plus/auth-strategies/WebAuth.js
@@ -1,7 +1,7 @@
 
 'use strict'
 
-const cst = require('../../../constants.js');
+const cst = require('../../../../constants.js');
 
 const AuthStrategy = require('@pm2/js-api/src/auth_strategies/strategy')
 const http = require('http')
