+++ lib/Client.js
@@ -168,7 +168,7 @@ Client.prototype.initFileStructure = function (opts) {
   }
 
   if (!process.env.PM2_PROGRAMMATIC && !fs.existsSync(path.join(opts.PM2_HOME, 'touch'))) {
-    var dt = fs.readFileSync(path.join(__dirname, opts.KEYMETRICS_BANNER));
+    var dt = fs.readFileSync(path.join(__dirname, opts.PM2_BANNER));
     console.log(dt.toString());
     try {
       fs.writeFileSync(path.join(opts.PM2_HOME, 'touch'), Date.now());
